Fix: api hook 분리
