+++ src/pages/sinitto/call-back/detail/api/accept-call-back.api.ts
@@ -1,0 +1,9 @@
+import { fetchInstance } from '@/shared/api/instance';
+
+const getacceptCallbackPath = (callbackId: number) =>
+  `/api/callbacks/accept/${callbackId}`;
+
+export const acceptCallback = async (callbackId: number) => {
+  const response = await fetchInstance.put(getacceptCallbackPath(callbackId));
+  return response.data;
+};
