+++ src/pages/sinitto/call-back/detail/api/hooks/useGetAccepted.ts
@@ -1,13 +1,5 @@
-import { fetchInstance } from '@/shared/api/instance';
-import type { CallbackResponse } from '@/shared/types';
+import { getAccepted } from '../get-accepted.api';
 import { useQuery } from '@tanstack/react-query';
-
-const getAcceptedPath = () => `/api/callbacks/sinitto/accepted`;
-
-const getAccepted = async () => {
-  const response = await fetchInstance.get<CallbackResponse>(getAcceptedPath());
-  return response.data;
-};
 
 export const useGetAccepted = () => {
   return useQuery({
