+++ src/pages/sinitto/call-back/detail/api/index.ts
@@ -1,0 +1,4 @@
+export { acceptCallback } from './accept-call-back.api';
+export { CancelCallback } from './cancel-call-back.api';
+export { CompleteCallback } from './complete-call-back.api';
+export { getAccepted } from './get-accepted.api';
