+++ src/pages/sinitto/call-back/list/api/index.ts
@@ -1,0 +1,1 @@
+export { getCallbacks } from './get-call-backs.api';
