+++ src/pages/sinitto/guide-line/api/index.ts
@@ -1,0 +1,1 @@
+export { getGuidelines } from './get-guideline.api';
